main.py
import krakenex
from pykrakenapi import KrakenAPI
import pandas as pd
import numpy as np
import time
from dotenv import load_dotenv
import os
import logging

load_dotenv()  # take environment variables from .env.


# Initialize Kraken API
API_KEY = os.getenv("API_KEY")
API_SECRET = os.getenv("API_SECRET")

api = krakenex.API(key=API_KEY, secret=API_SECRET)
kraken = KrakenAPI(api)

def calculate_rsi(data, period=14):
    # Implement RSI calculation
    pass

def buy_bitcoin(amount):
    # Use kraken API to place a buy order
    pass

def main():
    while True:
        # Fetch historical data
        ohlc, last = kraken.get_ohlc_data('XXBTZUSD', interval=60)
        
        # Calculate RSI
        rsi = calculate_rsi(ohlc['close'])
        
        # Check RSI condition
        if rsi[-1] < 20:
            buy_bitcoin(amount=0.01) # Specify your amount
        
        # Sleep for a certain amount of time before next check
        time.sleep(3600) # 1 hour

if __name__ == "__main__":
    main()
