requirements.txt
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
krakenex==2.2.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
pykrakenapi==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
